refactor

+++ SezzUI/Core/Animator/ColorAnimation.cs
@@ -10,7 +10,7 @@ namespace SezzUI.Animator
 
 		public ColorAnimation(Vector4 from, Vector4 to, uint duration = 0, uint delayStart = 0, uint delayEnd = 0)
 		{
-			_duration = (uint)Math.Max(50f, duration);
+			_duration = (uint) Math.Max(50f, duration);
 			_delayStart = delayStart;
 			_delayEnd = delayEnd;
 
@@ -23,14 +23,14 @@ namespace SezzUI.Animator
 			if (_isPlaying && _ticksStart != null)
 			{
 				int ticksNow = Environment.TickCount;
-				int timeElapsed = ticksNow - (int)_ticksStart;
+				int timeElapsed = ticksNow - (int) _ticksStart;
 
 				if (timeElapsed > StartDelay && timeElapsed <= StartDelay + Duration)
 				{
-					int timeElapsedAnimating = timeElapsed - (int)StartDelay;
+					int timeElapsedAnimating = timeElapsed - (int) StartDelay;
 
 					Vector4 range = ColorTo - ColorFrom;
-					float progress = Math.Min(1, Math.Max(0, (float)timeElapsedAnimating / (float)Duration));
+					float progress = Math.Min(1, Math.Max(0, timeElapsedAnimating / (float) Duration));
 					if (Data != null)
 					{
 						Data.Color = ColorFrom + range * progress;
@@ -44,6 +44,5 @@ namespace SezzUI.Animator
 				}
 			}
 		}
-
 	}
 }
